pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "athora_faq"
version = "0.1.0"
description = "Air-gapped RAG-based FAQ assistant for Athora products"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.31",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.7",
    "transformers>=4.40",
    "torch>=2.2",
    "sentence-transformers>=2.2",
    "streamlit>=1.38",
    "spacy>=3.7",
    "python-dotenv>=1.0",
    "pypdf>=4.0"
]

[tool.setuptools.packages.find]
where = ["src"]
